Don't delete files when discarding the post edit

+++ src/app/post/create-update-post/create-update-post.component.ts
@@ -82,7 +82,7 @@ export class CreateUpdatePostComponent implements OnInit {
   }
 
   ngOnDestroy() {
-    if (this.fileUrls.length > 0 && !this.postWasPosted) {
+    if (this.fileUrls.length > 0 && !this.postWasPosted && !this.isUpdating) {
       this.fileUrls.forEach(fileUrl => {
         this.fileService.removeFile(fileUrl);
       });
